+++ src/APIs/shoppingCart/addShoppingCart.ts
@@ -1,41 +1,12 @@
 import { CartItem } from "../../types/product.type";
-import { baseUrl, credentials } from "./apiConfig";
-import getShoppingCart from "./getShoppingCart";
+import fetchShoppingCart from "./fetchShoppingCart";
 import { ShoppingCartResponse } from "./types";
 
 async function addShoppingCart({
   endpoint,
   requestBody,
 }: ShoppingCartResponse): Promise<CartItem[]> {
-  try {
-    const response = await fetch(`${baseUrl}${endpoint}`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Basic ${credentials}`,
-      },
-      body: JSON.stringify(requestBody),
-    });
-
-    if (!response.ok) {
-      throw new Error("Network response was not ok");
-    }
-
-    const params = {
-      page: "0",
-      size: "50",
-    };
-
-    const query = new URLSearchParams(params).toString();
-
-    const responseDate = await getShoppingCart({
-      endpoint: `${endpoint}?${query}`,
-    });
-
-    return responseDate;
-  } catch (error) {
-    throw new Error("Error fetching products:" + error);
-  }
+  return fetchShoppingCart({ method: "POST", endpoint, requestBody });
 }
 
 export default addShoppingCart;
