+++ src/APIs/shoppingCart/deleteShoppingCart.ts
@@ -1,39 +1,12 @@
-import { CartItem, ShoppingCartResponse } from "../../types/product.type";
-import { baseUrl, credentials } from "./apiConfig";
-import getShoppingCart from "./getShoppingCart";
+import { CartItem } from "../../types/product.type";
+import fetchShoppingCart from "./fetchShoppingCart";
+import { ShoppingCartResponse } from "./types";
 
 async function deleteShoppingCart({
   endpoint,
   cartItemId,
 }: ShoppingCartResponse): Promise<CartItem[]> {
-  try {
-    const response = await fetch(`${baseUrl}${endpoint}/${cartItemId}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Basic ${credentials}`,
-      },
-    });
-
-    if (!response.ok) {
-      throw new Error("Network response was not ok");
-    }
-
-    const params = {
-      page: "0",
-      size: "50",
-    };
-
-    const query = new URLSearchParams(params).toString();
-
-    const responseDate = await getShoppingCart({
-      endpoint: `${endpoint}?${query}`,
-    });
-
-    return responseDate;
-  } catch (error) {
-    throw new Error("Error fetching products:" + error);
-  }
+  return fetchShoppingCart({ method: "DELETE", endpoint, cartItemId });
 }
 
 export default deleteShoppingCart;
